copy propagation tests

+++ tests/chapter_19/copy_propagation/all_types/dont_propagate/dont_propagate_addr_of.c
@@ -1,7 +1,7 @@
-int main(void)
-{
-    int x = 1;
-    int y = 2;
-    x = y;
-    return &x == &y; // don't rewrite as &y == &y
+/* Test that we don't propagate copies into AddrOf instructions */
+int main(void) {
+    long x = 1;
+    long y = 2;
+    x = y;            // gen x = y
+    return &x == &y;  // don't rewrite as &y == &y
 }
